+++ example/lib/example.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/widgets.dart';
 
 import 'font_awesome/regular.dart';
+import 'font_awesome/sharp_solid.dart';
 import 'font_awesome/solid.dart';
 
 const IconData icon = FASolid.faX;
 const IconData icon2 = FARegular.faMinus;
+const IconData icon3 = FASharpSolid.faAlbumCollectionCirclePlus;